backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
unstructured[pdf]
rank_bm25
